+++ scripts/ac5e-helpers.mjs
@@ -1,48 +1,196 @@
- import Constants from './ac5e-constants.mjs';
+import Constants from './ac5e-constants.mjs';
 import Settings from './ac5e-settings.mjs';
 
 const settings = new Settings();
 
 /**
+ * Foundry v12 updated.
  * Gets the minimum distance between two tokens,
- * evaluating all grid spaces they occupy, by Zhell.
+ * evaluating all grid spaces they occupy, based in Illandril's work 
+ * updated by thatlonelybugbear for 3D and tailored to AC5e needs!.
  */
 export function _getDistance(tokenA, tokenB) {
-	const A = _getAllTokenGridSpaces(tokenA.document);
-	const B = _getAllTokenGridSpaces(tokenB.document);
-	const rays = A.flatMap((a) => {
-		return B.map((b) => {
-			return { ray: new Ray(a, b) };
-		});
-	});
-	const dist = canvas.scene.grid.distance; // 5ft.
-	const distances = canvas.grid
-		.measureDistances(rays, {
-			gridSpaces: true,
-		})
-		.map((d) => Math.round(d / dist) * dist);
-	const eles = [tokenA, tokenB].map((t) => t.document.elevation);
-	const elevationDiff = Math.abs(eles[0] - eles[1]);
-	return Math.max(Math.min(...distances), elevationDiff);
-}
+    if (typeof tokenA === 'string' && !tokenA.includes('.')) tokenA = canvas.tokens.get(tokenA);
+    else if (typeof tokenA === 'string' && tokenA.includes('.')) tokenA = fromUuidSync(tokenA)?.object;
+    if (typeof tokenB === 'string' && !tokenB.includes('.')) tokenB = canvas.tokens.get(tokenB);
+    else if (typeof tokenB === 'string' && tokenB.includes('.')) tokenB = fromUuidSync(tokenB)?.object;
+    if (!tokenA || !tokenB) return false;
+    const PointsAndCenter = {
+        points: [],
+        trueCenternt: {},
+    };
 
-/**
- * Get the upper left corners of all grid spaces a token document occupies.
- */
-export function _getAllTokenGridSpaces(tokenDoc) {
-	const { width, height, x, y } = tokenDoc;
-	if (width <= 1 && height <= 1) return [{ x, y }];
-	const centers = [];
-	const grid = canvas.grid.size;
-	for (let a = 0; a < width; a++) {
-		for (let b = 0; b < height; b++) {
-			centers.push({
-				x: x + a * grid,
-				y: y + b * grid,
-			});
-		}
-	}
-	return centers;
+    const getPolygon = (grid /*: foundry.grid.BaseGrid*/, token /*: Token*/) => {
+        let poly; // PIXI.Polygon;
+        if (token.shape instanceof PIXI.Circle) {
+            poly = token.shape.toPolygon({ density: (token.shape.radius * 8) / grid.size });
+        } else if (token.shape instanceof PIXI.Rectangle) {
+            poly = token.shape.toPolygon();
+        } else {
+            poly = token.shape;
+        }
+
+        return new PIXI.Polygon(poly.points.map((point, i) => point + (i % 2 ? token.bounds.top : token.bounds.left)));
+    };
+
+    const getPointsAndCenter = (grid, shape) => {
+        const points = [];
+        for (let i = 0; i < shape.points.length; i += 2) {
+            const x = shape.points[i];
+            const y = shape.points[i + 1];
+            points.push({ x, y });
+
+            const nextX = shape.points[i + 2] ?? shape.points[0];
+            const nextY = shape.points[i + 3] ?? shape.points[1];
+            const d = Math.sqrt((x - nextX) ** 2 + (y - nextY) ** 2);
+            const steps = Math.ceil((d * 2) / grid.size);
+
+            for (let step = 1; step < steps; step++) {
+                points.push({ x: ((nextX - x) / steps) * step + x, y: ((nextY - y) / steps) * step + y });
+            }
+        }
+
+        return {
+            points: points,
+            trueCenter: shape.getBounds().center,
+        };
+    };
+
+    const getPoints = (grid /*: foundry.grid.BaseGrid*/, poly /*: PIXI.Polygon*/) => {
+        const bounds = poly.getBounds();
+        const pointsToMeasure = [bounds.center];
+
+        // If either dimension is one grid space long or less, just use the center point for measurements
+        // Otherwise, we use the center of the grid spaces along the token's perimeter
+        const forcedX = bounds.width <= grid.sizeX ? bounds.center.x : null;
+        const forcedY = bounds.height <= grid.sizeY ? bounds.center.x : null;
+
+        if (typeof forcedX !== 'number' || typeof forcedY !== 'number') {
+            const { points, trueCenter } = getPointsAndCenter(grid, poly);
+            for (const point of points) {
+                const x = (point.x - trueCenter.x) * 0.99 + trueCenter.x;
+                const y = (point.y - trueCenter.y) * 0.99 + trueCenter.y;
+                const pointToMeasure = grid.getCenterPoint({ x, y });
+                pointToMeasure.x = forcedX ?? pointToMeasure.x;
+                pointToMeasure.y = forcedY ?? pointToMeasure.y;
+                if (!pointsToMeasure.some((priorPoint) => priorPoint.x === pointToMeasure.x && priorPoint.y === pointToMeasure.y)) {
+                    pointsToMeasure.push(pointToMeasure);
+                }
+            }
+        }
+        return pointsToMeasure;
+    };
+
+    const squareDistance = (pointA /*: Point*/, pointB /*: Point*/) => (pointA.x - pointB.x) ** 2 + (pointA.y - pointB.y) ** 2;
+
+    const getComparisonPoints = (grid /*: foundry.grid.BaseGrid*/, token /*: Token*/, other /*: Token*/) => {
+        const polyA = getPolygon(grid, token);
+        const polyB = getPolygon(grid, other);
+
+        const pointsA = getPoints(grid, polyA);
+        const pointsB = getPoints(grid, polyB);
+        const containedPoint = pointsA.find((point) => polyB.contains(point.x, point.y)) ?? pointsB.find((point) => polyA.contains(point.x, point.y));
+        if (containedPoint) {
+            // A contains B or B contains A... so ensure the distance is 0
+            return [containedPoint, containedPoint];
+        }
+
+        let closestPointA = token.center;
+        let closestPointB = other.center;
+        let closestD2 = squareDistance(closestPointA, closestPointB);
+        for (const pointA of pointsA) {
+            for (const pointB of pointsB) {
+                const d2 = squareDistance(pointA, pointB);
+                if (d2 < closestD2) {
+                    closestD2 = d2;
+                    closestPointA = pointA;
+                    closestPointB = pointB;
+                }
+            }
+        }
+        return [closestPointA, closestPointB];
+    };
+    const calculateDistanceWithUnits = (scene, grid, token, other) => {
+        let totalDistance = 0;
+        let { distance, diagonals, spaces } = grid.measurePath(getComparisonPoints(grid, token, other));
+
+        if (canvas.grid.isSquare) {
+            token.z0 = token.document.elevation / grid.distance;
+            token.z1 = token.z0 + Math.min(token.document.width | 0, token.document.height | 0);
+            other.z0 = other.document.elevation / grid.distance;
+            other.z1 = other.z0 + Math.min(other.document.width | 0, other.document.height | 0);
+
+            let dz = other.z0 >= token.z1 ? other.z0 - token.z1 + 1 : token.z0 >= other.z1 ? token.z0 - other.z1 + 1 : 0;
+
+            if (!dz) {
+                totalDistance = distance;
+            } else {
+                const XY = { diagonals, illegal: spaces };
+                const Z = { illegal: dz, diagonals: Math.min(XY.illegal, dz) };
+                Z.diagonalsXYZ = Math.min(XY.diagonals, Z.diagonals);
+                Z.diagonalsXZ_YZ = Z.diagonals - Z.diagonalsXYZ;
+                XY.moves = spaces - (XY.diagonals + Z.diagonalsXZ_YZ);
+                Z.moves = dz - Z.diagonals;
+                const overallDiagonals = Math.max(XY.diagonals, Z.diagonals);
+
+                switch (grid.diagonals) {
+                    case CONST.GRID_DIAGONALS.EQUIDISTANT:
+                        totalDistance = XY.moves + Z.moves + overallDiagonals;
+                        break;
+
+                    case CONST.GRID_DIAGONALS.ALTERNATING_1:
+                        for (let i = 1; i <= overallDiagonals; i++) {
+                            totalDistance += i & 1 ? 1 : 2; // Odd/even check with bitwise
+                        }
+                        totalDistance += XY.moves + Z.moves;
+                        break;
+
+                    case CONST.GRID_DIAGONALS.ALTERNATING_2:
+                        for (let i = 1; i <= overallDiagonals; i++) {
+                            totalDistance += i & 1 ? 2 : 1; // Alternate between 2 and 1
+                        }
+                        totalDistance += XY.moves + Z.moves;
+                        break;
+
+                    case CONST.GRID_DIAGONALS.ILLEGAL:
+                        totalDistance = XY.illegal + Z.illegal;
+                        break;
+
+                    case CONST.GRID_DIAGONALS.EXACT:
+                        totalDistance = XY.moves + Z.moves + (overallDiagonals - Z.diagonalsXYZ) * Math.sqrt(2) + Z.diagonalsXYZ * Math.sqrt(3);
+                        break;
+
+                    case CONST.GRID_DIAGONALS.APPROXIMATE:
+                        totalDistance = XY.moves + Z.moves + overallDiagonals * 1.5;
+                        break;
+
+                    case CONST.GRID_DIAGONALS.RECTILINEAR:
+                        totalDistance = XY.moves + Z.moves + overallDiagonals * 2;
+                        break;
+
+                    default:
+                        throw new Error(`Unknown diagonal rule: ${grid.diagonals}`);
+                }
+
+                totalDistance *= grid.distance;
+            }
+        } else {
+            token.z0 = token.document.elevation;
+            token.z1 = token.z0 + Math.min(token.document.width * grid.distance, token.document.height * grid.distance);
+            other.z0 = other.document.elevation;
+            other.z1 = other.z0 + Math.min(other.document.width * grid.distance, other.document.height * grid.distance);
+            let dz = other.z0 > token.z1 ? other.z0 - token.z1 + grid.distance : token.z0 > other.z1 ? token.z0 - other.z1 + grid.distance : 0;
+            totalDistance = dz ? Math.sqrt(distance * distance + dz * dz) : distance;
+        }
+
+        return {
+            value: totalDistance,
+            units: scene.grid.units,
+        };
+    };
+    const result = calculateDistanceWithUnits(canvas.scene, canvas.grid, tokenA, tokenB).value;
+    if (settings.debug) console.log(`${Constants.MODULE.NAME.SHORT} - getDistane():`, { sourceId: token.id, targetId: other.id, result });
+    return (result * 100 | 0) / 100;
 }
 
 export function _i18nConditions(name) {
@@ -62,18 +210,11 @@ export function _hasStatuses(actor, statuses) {
 	const exhaustionNumberedStatus = statuses.find((s) => endsWithNumber(s));
 	if (exhaustionNumberedStatus) {
 		statuses = statuses.filter((s) => !endsWithNumber(s));
-		if (
-			_getExhaustionLevel(
-				actor,
-				exhaustionNumberedStatus.split('exhaustion')[1]
-			)
-		)
+		if (_getExhaustionLevel(actor, exhaustionNumberedStatus.split('exhaustion')[1]))
 			return [...actor.statuses]
 				.filter((s) => statuses.includes(s))
 				.map((el) => _i18nConditions(el.capitalize()))
-				.concat(
-					`${_i18nConditions('Exhaustion')} ${_getExhaustionLevel(actor)}`
-				)
+				.concat(`${_i18nConditions('Exhaustion')} ${_getExhaustionLevel(actor)}`)
 				.sort();
 	}
 	return [...actor.statuses]
@@ -89,24 +230,13 @@ export function _hasAppliedEffects(actor) {
 export function _getExhaustionLevel(actor, min = undefined, max = undefined) {
 	if (!actor) return false;
 	let exhaustionLevel = '';
-	const hasExhaustion =
-		actor.statuses.has('exhaustion') ||
-		actor.flags?.['automated-conditions-5e']?.statuses;
+	const hasExhaustion = actor.statuses.has('exhaustion') || actor.flags?.['automated-conditions-5e']?.statuses;
 	if (hasExhaustion) exhaustionLevel = actor.system.attributes.exhaustion;
 	return min ? min <= exhaustionLevel : exhaustionLevel;
 }
 
 export function _calcAdvantageMode(ac5eConfig, config) {
-	config.fastForward = config.fastForward
-		? config.fastForward
-		: ac5eConfig.roller == 'Core'
-		? config.event.shiftKey ||
-		  config.event.altKey ||
-		  config.event.metaKey ||
-		  config.event.ctrlKey
-		: ac5eConfig.roller == 'RSR'
-		? ac5eConfig.rsrOverrideFF
-		: false;
+	config.fastForward = config.fastForward ? config.fastForward : ac5eConfig.roller == 'Core' ? config.event.shiftKey || config.event.altKey || config.event.metaKey || config.event.ctrlKey : ac5eConfig.roller == 'RSR' ? ac5eConfig.rsrOverrideFF : false;
 	if (ac5eConfig.roller == 'Core')
 		foundry.utils.mergeObject(config.event, {
 			altKey: false,
@@ -123,16 +253,8 @@ export function _calcAdvantageMode(ac5eConfig, config) {
 		if (ac5eConfig.preAC5eConfig.disKey) return (config.disadvantage = true);
 		if (ac5eConfig.preAC5eConfig.critKey) return (config.critical = true);
 	}
-	if (
-		ac5eConfig.advantage.source?.length ||
-		ac5eConfig.advantage.target?.length
-	)
-		config.advantage = true;
-	if (
-		ac5eConfig.disadvantage.source?.length ||
-		ac5eConfig.disadvantage.target?.length
-	)
-		config.disadvantage = true;
+	if (ac5eConfig.advantage.source?.length || ac5eConfig.advantage.target?.length) config.advantage = true;
+	if (ac5eConfig.disadvantage.source?.length || ac5eConfig.disadvantage.target?.length) config.disadvantage = true;
 	if (ac5eConfig.advantage.length) config.advantage = true;
 	if (ac5eConfig.disadvantage.length) config.disadvantage = true;
 	if (config.advantage === true && config.disadvantage === true) {
@@ -162,11 +284,7 @@ export function _findNearby(
 	includeToken = true //includes or exclude source token
 ) {
 	if (!canvas || !canvas.tokens?.placeables) return false;
-	const validTokens = canvas.tokens.placeables.filter(
-		(placeable) =>
-			_dispositionCheck(token, placeable, disposition) &&
-			_getDistance(token, placeable) <= radius
-	);
+	const validTokens = canvas.tokens.placeables.filter((placeable) => _dispositionCheck(token, placeable, disposition) && _getDistance(token, placeable) <= radius);
 	if (lengthTest && includeToken) return validTokens.length >= lengthTest;
 	if (lengthTest && !includeToken) return validTokens.length > lengthTest;
 	if (includeToken) return validTokens;
@@ -178,33 +296,14 @@ export function _autoArmor(actor) {
 	const hasArmor = actor.armor;
 	const hasShield = actor.shield;
 	return {
-		hasStealthDisadvantage: 
-			hasArmor?.system.properties.has('stealthDisadvantage')
-			? 'Armor'
-			: hasShield?.system.properties.has('stealthDisadvantage')
-			? 'EquipmentShield'
-			: actor.itemTypes.equipment.some(
-					(item) =>
-						item.system.equipped &&
-						item.system.properties.has('stealthDisadvantage')
-			  )
-			? 'AC5E.Equipment'
-			: false,
-		notProficient:
-			!!hasArmor && !hasArmor.system.proficient && !hasArmor.system.prof.multiplier
-			? 'Armor' 
-			: !!hasShield && !hasShield.system.proficient && !hasShield.system.prof.multiplier
-			? 'EquipmentShield'
-			: false,
+		hasStealthDisadvantage: hasArmor?.system.properties.has('stealthDisadvantage') ? 'Armor' : hasShield?.system.properties.has('stealthDisadvantage') ? 'EquipmentShield' : actor.itemTypes.equipment.some((item) => item.system.equipped && item.system.properties.has('stealthDisadvantage')) ? 'AC5E.Equipment' : false,
+		notProficient: !!hasArmor && !hasArmor.system.proficient && !hasArmor.system.prof.multiplier ? 'Armor' : !!hasShield && !hasShield.system.proficient && !hasShield.system.prof.multiplier ? 'EquipmentShield' : false,
 	};
 }
 
 export function _autoEncumbrance(actor, abilityId) {
 	if (!settings.autoEncumbrance) return null;
-	return (
-		['con', 'dex', 'str'].includes(abilityId) &&
-		_hasStatuses(actor, 'heavilyEncumbered').length
-	);
+	return ['con', 'dex', 'str'].includes(abilityId) && _hasStatuses(actor, 'heavilyEncumbered').length;
 }
 
 export function _autoRanged(item, token, target) {
@@ -229,11 +328,7 @@ export function _autoRanged(item, token, target) {
 
 
 export function _hasItem(actor, itemName) {
-	return actor?.items.some((item) =>
-		item?.name
-			.toLocaleLowerCase()
-			.includes(_localize(itemName).toLocaleLowerCase())
-	);
+	return actor?.items.some((item) => item?.name.toLocaleLowerCase().includes(_localize(itemName).toLocaleLowerCase()));
 }
 
 export function _systemCheck(testVersion) {
@@ -241,82 +336,35 @@ export function _systemCheck(testVersion) {
 }
 
 export function _getTooltip(ac5eConfig) {
-	let tooltip = settings.showNameTooltips
-		? '<center><strong>Automated Conditions 5e</strong></center><hr>'
-		: '';
-	if (ac5eConfig.critical.length)
-		tooltip = tooltip.concat(
-			`<span style="display: block; text-align: left;">${_localize(
-				'Critical'
-			)}: ${ac5eConfig.critical.join(', ')}</span>`
-		);
+	let tooltip = settings.showNameTooltips ? '<center><strong>Automated Conditions 5e</strong></center><hr>' : '';
+	if (ac5eConfig.critical.length) tooltip = tooltip.concat(`<span style="display: block; text-align: left;">${_localize('Critical')}: ${ac5eConfig.critical.join(', ')}</span>`);
 	if (ac5eConfig.advantage?.length) {
 		if (tooltip.includes(':')) tooltip = tooltip.concat('<br>');
-		tooltip = tooltip.concat(
-			`<span style="display: block; text-align: left;">${_localize(
-				'Advantage'
-			)}: ${ac5eConfig.advantage.join(', ')}</span>`
-		);
+		tooltip = tooltip.concat(`<span style="display: block; text-align: left;">${_localize('Advantage')}: ${ac5eConfig.advantage.join(', ')}</span>`);
 	}
 	if (ac5eConfig.disadvantage?.length) {
 		if (tooltip.includes(':')) tooltip = tooltip.concat('<br>');
-		tooltip = tooltip.concat(
-			`<span style="display: block; text-align: left;">${_localize(
-				'Disadvantage'
-			)}: ${ac5eConfig.disadvantage.join(', ')}</span>`
-		);
+		tooltip = tooltip.concat(`<span style="display: block; text-align: left;">${_localize('Disadvantage')}: ${ac5eConfig.disadvantage.join(', ')}</span>`);
 	}
 	if (ac5eConfig.fail?.length) {
 		if (tooltip.includes(':')) tooltip = tooltip.concat('<br>');
-		tooltip = tooltip.concat(
-			`<span style="display: block; text-align: left;">${_localize(
-				'AC5E.Fail'
-			)}: ${ac5eConfig.fail.join(', ')}</span>`
-		);
+		tooltip = tooltip.concat(`<span style="display: block; text-align: left;">${_localize('AC5E.Fail')}: ${ac5eConfig.fail.join(', ')}</span>`);
 	}
 	if (ac5eConfig.advantage?.source?.length) {
 		if (tooltip.includes(':')) tooltip = tooltip.concat('<br>');
-		tooltip = tooltip.concat(
-			`<span style="display: block; text-align: left;">Attacker ${_localize(
-				'Advantage'
-			)
-				.substring(0, 3)
-				.toLocaleLowerCase()}: ${ac5eConfig.advantage.source.join(', ')}</span>`
-		);
+		tooltip = tooltip.concat(`<span style="display: block; text-align: left;">Attacker ${_localize('Advantage').substring(0, 3).toLocaleLowerCase()}: ${ac5eConfig.advantage.source.join(', ')}</span>`);
 	}
 	if (ac5eConfig.advantage?.target?.length) {
 		if (tooltip.includes(':')) tooltip = tooltip.concat('<br>');
-		tooltip = tooltip.concat(
-			`<span style="display: block; text-align: left;">${_localize(
-				'Target'
-			)} grants ${_localize('Advantage')
-				.substring(0, 3)
-				.toLocaleLowerCase()}: ${ac5eConfig.advantage.target.join(', ')}</span>`
-		);
+		tooltip = tooltip.concat(`<span style="display: block; text-align: left;">${_localize('Target')} grants ${_localize('Advantage').substring(0, 3).toLocaleLowerCase()}: ${ac5eConfig.advantage.target.join(', ')}</span>`);
 	}
 	if (ac5eConfig.disadvantage?.source?.length) {
 		if (tooltip.includes(':')) tooltip = tooltip.concat('<br>');
-		tooltip = tooltip.concat(
-			`<span style="display: block; text-align: left;">Attacker ${_localize(
-				'Disadvantage'
-			)
-				.substring(0, 3)
-				.toLocaleLowerCase()}: ${ac5eConfig.disadvantage.source.join(
-				', '
-			)}</span>`
-		);
+		tooltip = tooltip.concat(`<span style="display: block; text-align: left;">Attacker ${_localize('Disadvantage').substring(0, 3).toLocaleLowerCase()}: ${ac5eConfig.disadvantage.source.join(', ')}</span>`);
 	}
 	if (ac5eConfig.disadvantage?.target?.length) {
 		if (tooltip.includes(':')) tooltip = tooltip.concat('<br>');
-		tooltip = tooltip.concat(
-			`<span style="display: block; text-align: left;">${_localize(
-				'Target'
-			)} grants ${_localize('Disadvantage')
-				.substring(0, 3)
-				.toLocaleLowerCase()}: ${ac5eConfig.disadvantage.target.join(
-				', '
-			)}</span>`
-		);
+		tooltip = tooltip.concat(`<span style="display: block; text-align: left;">${_localize('Target')} grants ${_localize('Disadvantage').substring(0, 3).toLocaleLowerCase()}: ${ac5eConfig.disadvantage.target.join(', ')}</span>`);
 	}
 	if (!tooltip.includes(':')) return null;
 	else return tooltip;
@@ -327,18 +375,10 @@ export function _getConfig(config, hookType, tokenId, targetId) {
 	const existingAC5e = config?.dialogOptions?.['automated-conditions-5e'];
 	if (!!existingAC5e && existingAC5e.hookType == 'item') {
 		existingAC5e.hookType = existingAC5e.hookType + hookType.capitalize();
-		if (settings.debug)
-			console.warn('ac5e helpers.getConfig preExistingAC5e:', preExistingAC5e);
+		if (settings.debug) console.warn('ac5e helpers.getConfig preExistingAC5e:', preExistingAC5e);
 		return existingAC5e;
 	}
-	if (settings.debug)
-		console.log(
-			config.advantage,
-			config.disadvantage,
-			config.critical,
-			config.fastForward,
-			hookType
-		);
+	if (settings.debug) console.log(config.advantage, config.disadvantage, config.critical, config.fastForward, hookType);
 	let moduleID = 'Core';
 	let advKey,
 		disKey,
@@ -363,85 +403,34 @@ export function _getConfig(config, hookType, tokenId, targetId) {
 			if (rsrHookType == 'attack') rsrHookType = 'item';
 			if (rsrHookType == 'conc') rsrHookType = 'ability';
 			rsrHookType = rsrHookType.capitalize();
-			advKey = !getRsrSetting(`enable${rsrHookType}QuickRoll`)
-				? config.event?.altKey || config.event?.metaKey
-				: getRsrSetting('rollModifierMode') == 0
-				? config.event?.shiftKey
-				: config.event?.ctrlKey || config.event?.metaKey;
-			disKey = !getRsrSetting(`enable${rsrHookType}QuickRoll`)
-				? config.event?.ctrlKey
-				: getRsrSetting('rollModifierMode') == 0
-				? config.event?.ctrlKey || config.event?.metaKey
-				: config.event?.shiftKey;
-			rsrOverrideFF = getRsrSetting(`enable${rsrHookType}QuickRoll`)
-				? !config.event?.altKey
-				: config.event.shiftKey ||
-				  config.event.altKey ||
-				  config.event.metaKey ||
-				  config.event.ctrlKey;
+			advKey = !getRsrSetting(`enable${rsrHookType}QuickRoll`) ? config.event?.altKey || config.event?.metaKey : getRsrSetting('rollModifierMode') == 0 ? config.event?.shiftKey : config.event?.ctrlKey || config.event?.metaKey;
+			disKey = !getRsrSetting(`enable${rsrHookType}QuickRoll`) ? config.event?.ctrlKey : getRsrSetting('rollModifierMode') == 0 ? config.event?.ctrlKey || config.event?.metaKey : config.event?.shiftKey;
+			rsrOverrideFF = getRsrSetting(`enable${rsrHookType}QuickRoll`) ? !config.event?.altKey : config.event.shiftKey || config.event.altKey || config.event.metaKey || config.event.ctrlKey;
 		} else if (rsrHookType == 'damage') {
 			//to-do:check this
 			rsrHookType = 'Item';
-			critKey = !getRsrSetting(`enable${rsrHookType}QuickRoll`)
-				? config.event?.altKey || config.event?.metaKey
-				: getRsrSetting('rollModifierMode') == 0
-				? config.event?.shiftKey
-				: config.event?.ctrlKey || config.event?.metaKey;
-			rsrOverrideFF = getRsrSetting(`enable${rsrHookType}QuickRoll`)
-				? !config.event?.altKey
-				: config.event?.shiftKey;
+			critKey = !getRsrSetting(`enable${rsrHookType}QuickRoll`) ? config.event?.altKey || config.event?.metaKey : getRsrSetting('rollModifierMode') == 0 ? config.event?.shiftKey : config.event?.ctrlKey || config.event?.metaKey;
+			rsrOverrideFF = getRsrSetting(`enable${rsrHookType}QuickRoll`) ? !config.event?.altKey : config.event?.shiftKey;
 		}
 		moduleID = 'RSR';
 	} else {
 		//core system keys
-		if (hookType != 'damage')
-			advKey = config.event?.altKey || config.event?.metaKey;
+		if (hookType != 'damage') advKey = config.event?.altKey || config.event?.metaKey;
 		if (hookType != 'damage') disKey = config.event?.ctrlKey;
-		if (hookType == 'damage')
-			critKey = config.event?.altKey || config.event?.metaKey;
+		if (hookType == 'damage') critKey = config.event?.altKey || config.event?.metaKey;
 	}
-	if (settings.debug)
-		console.warn(
-			'helpers check Keys || ',
-			hookType,
-			advKey,
-			disKey,
-			critKey,
-			moduleID,
-			'keypressOverrides:',
-			settings.keypressOverrides
-		);
+	if (settings.debug) console.warn('helpers check Keys || ', hookType, advKey, disKey, critKey, moduleID, 'keypressOverrides:', settings.keypressOverrides);
 	if (advKey) advantage = [`${moduleID} (keyPress)`];
 	if (disKey) disadvantage = [`${moduleID} (keyPress)`];
-	if (critKey && ['damage', 'itemDamage'].includes(hookType))
-		critical = [`${moduleID} (keyPress)`];
-	if (config.advantage /*&& !settings.keypressOverrides*/)  //to-do: why was that here in the first place? Changed when added multi rollers compat?
+	if (critKey && ['damage', 'itemDamage'].includes(hookType)) critical = [`${moduleID} (keyPress)`];
+	if (config.advantage /*&& !settings.keypressOverrides*/)
+		//to-do: why was that here in the first place? Changed when added multi rollers compat?
 		advantage = advantage.concat(`${moduleID} (flags)`);
-	if (config.disadvantage /*&& !settings.keypressOverrides*/)
-		disadvantage = disadvantage.concat(`${moduleID} (flags)`);
-	if (config.critical === true /*&& !settings.keypressOverrides*/)
-		critical = critical.concat(`${moduleID} (flags)`);
+	if (config.disadvantage /*&& !settings.keypressOverrides*/) disadvantage = disadvantage.concat(`${moduleID} (flags)`);
+	if (config.critical === true /*&& !settings.keypressOverrides*/) critical = critical.concat(`${moduleID} (flags)`);
 	if (settings.debug) {
-		console.warn(
-			'_getConfig | advantage:',
-			advantage,
-			'disadvantage:',
-			disadvantage,
-			'critical:',
-			critical,
-			'hookType:',
-			hookType
-		);
-		console.warn(
-			'_getConfig keys | advKey:',
-			advKey,
-			'disKey:',
-			disKey,
-			'critKey:',
-			critKey,
-			'rsrOverrideFF:',
-			rsrOverrideFF
-		);
+		console.warn('_getConfig | advantage:', advantage, 'disadvantage:', disadvantage, 'critical:', critical, 'hookType:', hookType);
+		console.warn('_getConfig keys | advKey:', advKey, 'disKey:', disKey, 'critKey:', critKey, 'rsrOverrideFF:', rsrOverrideFF);
 	}
 	return {
 		hookType,
@@ -468,25 +457,119 @@ export function _setAC5eProperties(ac5eConfig, where) {
 		foundry.utils.mergeObject(where.dialogOptions, {
 			[`${Constants.MODULE_ID}`]: ac5eConfig,
 		});
-	else
-		foundry.utils.setProperty(
-			where,
-			`dialogOptions.${Constants.MODULE_ID}`,
-			ac5eConfig
-		);
+	else foundry.utils.setProperty(where, `dialogOptions.${Constants.MODULE_ID}`, ac5eConfig);
 	foundry.utils.mergeObject(where.dialogOptions, { classes: ['ac5e dialog'] });
 	if (where.messageData)
 		foundry.utils.mergeObject(where.messageData, {
 			[`flags.${Constants.MODULE_ID}`]: ac5eConfig,
 		});
-	else
-		foundry.utils.setProperty(
-			where,
-			`messageData.flags.${Constants.MODULE_ID}`,
-			ac5eConfig
-		);
+	else foundry.utils.setProperty(where, `messageData.flags.${Constants.MODULE_ID}`, ac5eConfig);
 }
 
 function activeModule(moduleID) {
 	return game.modules.get(moduleID)?.active;
+}
+
+export function _canSee(source, target) {
+	//if (game.modules.get('midi - qol')?.active) return MidiQOL.canSee(source, target);
+	if (!source || !target) {
+		if (settings.debug) console.warn('AC5e: No valid tokens for canSee test');
+		return false;
+	}
+	//any non-owned, non-selected tokens will have their vision not initialized.
+	if (!source.vision) _initializeVision(source);
+	if (!target.vision) _initializeVision(target);
+	const NON_SIGHT_CONSIDERED_SIGHT = ['blindsight'];
+	const detectionModes = CONFIG.Canvas.detectionModes;
+	const DetectionModeCONST = DetectionMode;
+	const sightDetectionModes = new Set(Object.keys(detectionModes).filter((d) => detectionModes[d].type === DetectionMode.DETECTION_TYPES.SIGHT || NON_SIGHT_CONSIDERED_SIGHT.includes[d])); // ['basicSight', 'seeInvisibility', 'seeAll']
+	if (source instanceof TokenDocument) source = source.object;
+	if (target instanceof TokenDocument) target = target.object;
+	if (target.document?.hidden) return false;
+	if (!source.hasSight) return true; //if no sight is enabled on the source, it can always see.
+
+	const matchedModes = new Set();
+	// Determine the array of offset points to test
+	const t = Math.min(target.w, target.h) / 4;
+	const targetPoint = target.center;
+	const offsets =
+		t > 0
+			? [
+					[0, 0],
+					[-t, -t],
+					[-t, t],
+					[t, t],
+					[t, -t],
+					[-t, 0],
+					[t, 0],
+					[0, -t],
+					[0, t],
+			  ]
+			: [[0, 0]];
+	const tests = offsets.map((o) => ({
+		point: new PIXI.Point(targetPoint.x + o[0], targetPoint.y + o[1]),
+		los: new Map(),
+	}));
+	const config = { tests, object: target };
+	const tokenDetectionModes = source.detectionModes;
+	// First test basic detection for light sources which specifically provide vision
+	const lightSources = foundry.utils.isNewerVersion(game.system.version, '12.0') ? canvas?.effects?.lightSources : canvas?.effects?.lightSources.values();
+	for (const lightSource of lightSources ?? []) {
+		if (!lightSource.active || lightSource.data.disabled) continue;
+		if (!lightSource.data.visibility) continue;
+		const result = lightSource.testVisibility(config);
+		if (result === true) matchedModes.add(detectionModes.lightPerception?.id ?? DetectionModeCONST.BASIC_MODE_ID);
+	}
+	const basic = tokenDetectionModes.find((m) => m.id === DetectionModeCONST.BASIC_MODE_ID);
+	if (basic) {
+		if (['basicSight', 'lightPerception', 'all'].some((mode) => sightDetectionModes.has(mode))) {
+			const result = source.vision ? detectionModes.basicSight?.testVisibility(source.vision, basic, config) : false;
+			if (result === true) matchedModes.add(detectionModes.lightPerception?.id ?? DetectionModeCONST.BASIC_MODE_ID);
+		}
+	}
+	for (const detectionMode of tokenDetectionModes) {
+		if (detectionMode.id === DetectionModeCONST.BASIC_MODE_ID) continue;
+		if (!detectionMode.enabled) continue;
+		const dm = sightDetectionModes[detectionMode.id];
+		if (sightDetectionModes.has('all') || sightDetectionModes.has(detectionMode.id)) {
+			const result = dm?.testVisibility(source.vision, detectionMode, config);
+			if (result === true) {
+				matchedModes.add(detectionMode.id);
+			}
+		}
+	}
+	if (settings.debug) console.warn(`${Constants.MODULE_SHORT_NAME} - _canSee()`, { sourceId: source?.id, targetId: target?.id, result: matchedModes });
+	return !!matchedModes.size;
+}
+
+function _initializeVision(token) {
+	const sightEnabled = token.document.sight.enabled;
+	token.document.sight.enabled = true;
+	token.document._prepareDetectionModes();
+	const sourceId = token.sourceId;
+	token.vision = new CONFIG.Canvas.visionSourceClass({ sourceId, object: token }); //v12 only
+	token.vision.initialize({
+		x: token.center.x,
+		y: token.center.y,
+		elevation: token.document.elevation,
+		radius: Math.clamp(token.sightRange, 0, canvas?.dimensions?.maxR ?? 0),
+		externalRadius: token.externalRadius,
+		angle: token.document.sight.angle,
+		contrast: token.document.sight.contrast,
+		saturation: token.document.sight.saturation,
+		brightness: token.document.sight.brightness,
+		attenuation: token.document.sight.attenuation,
+		rotation: token.document.rotation,
+		visionMode: token.document.sight.visionMode,
+		color: globalThis.Color.from(token.document.sight.color),
+		isPreview: !!token._original,
+		blinded: token.document.hasStatusEffect(CONFIG.specialStatusEffects.BLIND),
+	});
+	if (!token.vision.los) {
+		token.vision.shape = token.vision._createRestrictedPolygon();
+		token.vision.los = token.vision.shape;
+	}
+	token.vision.animated = false;
+	canvas?.effects?.visionSources.set(sourceId, token.vision);
+	return true;
 }